+++ examples/init/basic.ts
@@ -26,9 +26,8 @@ const flow = async (context) => {
 };
 
 init(flow, {
-    logger: {
-        mode: "text",
-        route: "console",
-    },
+    mode: "text",
+    route: "console",
+    modules: ["logger"],
 });